// react-module/src/openapi/models/iproduct-defs-with-storage-amount.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface IProductDefsWithStorageAmount
 */
export interface IProductDefsWithStorageAmount {
    /**
     * 
     * @type {number}
     * @memberof IProductDefsWithStorageAmount
     */
    id?: any;
    /**
     * 
     * @type {number}
     * @memberof IProductDefsWithStorageAmount
     */
    availableAmount?: any;
    /**
     * 
     * @type {string}
     * @memberof IProductDefsWithStorageAmount
     */
    warehouseName?: any;
    /**
     * 
     * @type {number}
     * @memberof IProductDefsWithStorageAmount
     */
    defId?: any;
    /**
     * 
     * @type {string}
     * @memberof IProductDefsWithStorageAmount
     */
    storageAreaCode?: any;
    /**
     * 
     * @type {string}
     * @memberof IProductDefsWithStorageAmount
     */
    defTitle?: any;
    /**
     * 
     * @type {number}
     * @memberof IProductDefsWithStorageAmount
     */
    storageAreaId?: any;
    /**
     * 
     * @type {number}
     * @memberof IProductDefsWithStorageAmount
     */
    totalAmount?: any;
}