// react-module/src/openapi/models/iproduct-defs-with-total-amount.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface IProductDefsWithTotalAmount
 */
export interface IProductDefsWithTotalAmount {
    /**
     * 
     * @type {number}
     * @memberof IProductDefsWithTotalAmount
     */
    id?: any;
    /**
     * 
     * @type {ILayerShape}
     * @memberof IProductDefsWithTotalAmount
     */
    layerShape?: any;
    /**
     * 
     * @type {number}
     * @memberof IProductDefsWithTotalAmount
     */
    availableAmount?: any;
    /**
     * 
     * @type {number}
     * @memberof IProductDefsWithTotalAmount
     */
    defId?: any;
    /**
     * 
     * @type {string}
     * @memberof IProductDefsWithTotalAmount
     */
    defTitle?: any;
    /**
     * 
     * @type {number}
     * @memberof IProductDefsWithTotalAmount
     */
    totalAmount?: any;
}